==> weather-mapping/1-collect.ipynb <==
# Weather Mapping - Data Collection

### Dependencies

# Dependencies
import os
import requests
import numpy as np
import pandas as pd
import datetime as dt
from citipy import citipy
from config import OWM_KEY

### Generate at least 1000 cities around the world randomly

# Generate random GCS coords
np.random.seed(0)
n_coords = 3500
lat, lon = np.random.uniform(-90, 90, n_coords), np.random.uniform(-180, 180, n_coords)
lat[:5], lon[:5]

# Find cities closest to the generated coords
cities = []
for i in range(n_coords):
    city = citipy.nearest_city(lat[i], lon[i])
    if city.city_name not in cities:
        cities.append(city.city_name)
    
len(cities)

### Get weather data for each city

# OpenWeatherMap URL for API calls
url = 'http://api.openweathermap.org/data/2.5/weather?units=imperial'
url += '&appid=' + OWM_KEY + '&q='

# Sample OWM response
response = requests.get(url + cities[0]).json()
response

# Store 
weather200, weather404 = [], []

# Print header
print('Starting data collection...\n')
sep = (' ' * 7) + '| ' # col separator
header = 'Num API requests | Cities collected | Cities not found'
print(header)
print('-' * len(header))

# Get weather for each city
for i, city in enumerate(cities):
    
    # Make request
    response = requests.get(url + city.replace(' ', '+')).json()
    
    # For 200 responses (weather was found for city)
    try: 
        
        # Main weather
        weather = {
            'City': response['name'],
            'Country': response['sys']['country'],
            'Latitude': response['coord']['lat'],
            'Longitude': response['coord']['lon'],
            'Time': dt.datetime.utcfromtimestamp(response['dt']).strftime('%Y-%m-%d %H:%M:%S'),
            'Description': response['weather'][0]['description'],
            'Max Temp': response['main']['temp_max'],
            'Humidity': response['main']['humidity'],
            'Wind Speed': response['wind']['speed'],
            'Cloudiness': response['clouds']['all']
        }
        
        # Add rain and snow
        for meas in ['Rain', 'Snow']:
            try:
                weather[meas] = response[mean.lower()]['1h']
            except:
                weather[meas] = 0
        
        # Add weather to 200 list
        weather200.append(weather)
    
    except: # for 404 responses (city was not found)
        weather404.append(city) # add city to 404 list
        
    # Print progress every 100 cities and on the last city
    if (i > 0 and (i + 1) % 100 == 0) or i == len(cities) - 1:
        print(f'{i + 1:10}', end=sep) # cities requested
        print(f'{len(weather200):10}', end=sep) # cities collected
        print(f'{len(weather404):10}') # cities not found
        
print('-' * len(header))
print('\nData collection complete.\n')
print('The following cities were not found with the OpenWeatherMap API:')
print(weather404)

### Convert data to dataframe

# Create dataframe for weather data
weather_df = pd.DataFrame(weather200)
weather_df.head(2)

# Data path
os.makedirs('data', exist_ok=True) # create data dir
data_path = os.path.join('data', 'weather.csv')
data_path

# Save data
weather_df.to_csv(data_path, index_label='Id')
pd.read_csv(data_path).info()
