==> weather-mapping/3-search.ipynb <==
# WeatherPy - Hotel Search

### Dependencies and data

# Dependencies
import os
import requests
import gmaps
import numpy as np
import pandas as pd
from config import GMAPS_KEY

# Configure API key
gmaps.configure(api_key=GMAPS_KEY)

# Weather data
weather_path = os.path.join('data', 'weather.csv')
weather_df = pd.read_csv(weather_path)
weather_df.head(2)

### Find hotels near destinations within the user's preferred weather

# Lowest and highest `Max Temp` in the data
min_mt, max_mt = weather_df['Max Temp'].min(), weather_df['Max Temp'].max()
min_mt, max_mt

def user_input(prompt, min_temp=min_mt, max_temp=max_mt, binary=False):
    
    """
    Take user input for weather preferences.
    
    Parameters
    ----------
    prompt : str
        Input prompt
    min_temp : float, optional
        Minimum temperature in data, by default {min_mt}
    max_temp : float, optional
        Maximum temperature in data, by default {max_mt}
    binary : bool, optional
        Whether the input should be "yes" or "no" only, by default False
        
    Returns
    -------
    float
        User input
    """
    
    # Init user input and loop condition to False
    i = b = False
    
    # Get user input
    if binary: # yes/no input
        while i not in ['yes', 'no']: # continue until input is either "yes" or "no"
            i = input(prompt) # user input
        i = 1 if i == 'yes' else 0 # convert input to num
    else: 
        prompt += f'\tEnter a number between {min_temp} and {max_temp}: ' 
        while not b: # loop condition
            try:
                i = float(input(prompt)) # get user input as num
            except:
                continue
            if min_temp <= i <= max_temp: # check if user input is within proper temperature range
                b = True # end loop
            
    # Numeric input
    return i
        
    
# Get preferred weather
min_temp = user_input('1. What is the lowest temperature you would like at your destination?\n')
max_temp = user_input('2. What is the highest temperature you would like at your destination?\n', min_temp=min_temp)
rain = user_input('3. Would you like it to rain at your destination? (yes/no) ', binary=True)
snow = user_input('4. Would you like it to snow at your destination? (yes/no) ', binary=True)
min_temp, max_temp, rain, snow # enter 70, 80, no, no

# Filter for cities with the temp preference
pref_cities = weather_df.loc[(weather_df['Max Temp'] >= min_temp) & 
                             (weather_df['Max Temp'] <= max_temp)]

# Filter rain preference
if rain:
    pref_cities = pref_cities.query('Rain > 0')
else:
    pref_cities = pref_cities.query('Rain == 0')

# Filter snow preference
if snow:
    pref_cities = pref_cities.query('Snow > 0')
else:
    pref_cities = pref_cities.query('Snow == 0')
    
pref_cities.shape

# New df for hotels
hotel_df = pref_cities.reset_index(drop=True)
hotel_df['Hotel'] = np.NaN # new col for hotels

# Base url and params for GMaps Places API
base_url = 'https://maps.googleapis.com/maps/api/place/nearbysearch/json'
payload = {
    'key': GMAPS_KEY,
    'type': 'lodging', # hotels
    'radius': 5000 # meters
}

hotel_df.head(2)

# Sample hotel response
payload['location'] = f"{hotel_df.loc[0, 'Latitude']},{hotel_df.loc[0, 'Longitude']}" # 1st city
response = requests.get(base_url, params=payload)
response.json()['results'][0]

# Indices of cities without results
hotel404 = []
print('Starting hotel requests...')

# Find nearest hotel for each city
for idx, row in hotel_df.iterrows():
    
    # Print progress
    if (idx > 0 and idx % 50 == 0) or idx == hotel_df.shape[0] - 1:
        print(f'Requested hotels for {idx} locations.')
        
    # Request hotels
    payload['location'] = f'{row["Latitude"]}, {row["Longitude"]}' # city coordinates
    try:
        response = requests.get(base_url, params=payload).json()
        hotel_df.loc[idx, 'Hotel'] = response['results'][0]['name']
    except:
        hotel404.append(idx)
        
print('Hotel requests complete.\n')
print('Hotels were not found for the following cities:')
print(len(hotel404), hotel_df.loc[hotel404, 'City'].tolist())

# Drop cities without hotel results
hotel_df.dropna(inplace=True)
hotel_df.shape

### Temperature heatmap

# For Jupyterlab
# !jupyter labextension install @jupyter-widgets/jupyterlab-manager
# !jupyter lab build

# Locations and measurements
locations = hotel_df[['Latitude', 'Longitude']]
weights = hotel_df['Max Temp'].apply(lambda t: max(t, 0))

# Popup info box
hotel_info = """
<dl>
<dt>Hotel</dt><dd>{Hotel}</dd>
<dt>City</dt><dd>{City}, {Country}</dd>
<dt>Weather</dt><dd>{Max Temp} F - {Description}</dd>
</dl>
"""

# Layers
heatmap = gmaps.heatmap_layer(locations, weights, dissipating=False, point_radius=3, max_intensity=300)
markers = gmaps.marker_layer(locations,
                             info_box_content=[hotel_info.format(**row) for _, row in hotel_df.iterrows()])

# Heatmap
fig = gmaps.figure(center=(0.0, 0.0), zoom_level=1.5)
fig.add_layer(heatmap)
fig.add_layer(markers)
fig

# Save hotel data
hotel_path = os.path.join('data', 'hotels.csv')
hotel_df.to_csv(hotel_path, index=False)
pd.read_csv(hotel_path).info()
